Sketch boot sequence in full
There may still be changes, I've not yet been able to find which
sequence order is important and which may be adjusted from the Linux
kernel order. (Beyond what is trivial from the fact that we should be at
least as ordered as if the mmu and cci was already initialized, and to
avoid data races).